添加部署环境工具
