+++ Workspace/Proxy/Dockerfile
@@ -1,0 +1,3 @@
+FROM nginx
+COPY nginx /etc/nginx
+COPY html  /usr/share/nginx/html
