+++ Workspace/Proxy/nginx/conf.d/default.conf
@@ -1,0 +1,80 @@
+server {
+    listen  80;
+    server_name localhost;
+      
+    location /docs/swagger {
+        proxy_pass http://dlog:9090;
+        proxy_set_header Host $proxy_host;
+        proxy_set_header Cookie $http_cookie;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header Via "nginx";
+    }
+    location /api {
+        proxy_pass http://dlog:9090;
+        proxy_set_header Host $proxy_host;
+        proxy_set_header Cookie $http_cookie;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header Via "nginx";
+    }
+
+    location / {
+        root   /usr/share/nginx/html;
+        index  index.html index.htm;
+    }
+
+    # rewrite ^(.*)$  https://$host$1 permanent; 
+}
+
+# server {
+    # listen 443 ssl;
+    # server_name  localhost;
+    # ssl                      on;
+    # ssl_certificate          /etc/nginx/certs/example.crt;
+    # ssl_certificate_key      /etc/nginx/certs/example.key;
+
+    # ssl_session_timeout  5m;
+
+    # ssl_ciphers HIGH:!aNULL:!MD5;
+    # ssl_protocols SSLv3 TLSv1 TLSv1.1 TLSv1.2;
+    # ssl_prefer_server_ciphers   on;
+
+    # location / {
+    #     root   /usr/share/nginx/html;
+    #     index  index.html index.htm;
+    # }
+
+    #error_page  404              /404.html;
+
+    # redirect server error pages to the static page /50x.html
+    #
+    # error_page   500 502 503 504  /50x.html;
+    # location = /50x.html {
+    #     root   /usr/share/nginx/html;
+    # }
+
+    # proxy the PHP scripts to Apache listening on 127.0.0.1:80
+    #
+    #location ~ \.php$ {
+    #    proxy_pass   http://127.0.0.1;
+    #}
+
+    # pass the PHP scripts to FastCGI server listening on 127.0.0.1:9000
+    #
+    #location ~ \.php$ {
+    #    root           html;
+    #    fastcgi_pass   127.0.0.1:9000;
+    #    fastcgi_index  index.php;
+    #    fastcgi_param  SCRIPT_FILENAME  /scripts$fastcgi_script_name;
+    #    include        fastcgi_params;
+    #}
+
+    # deny access to .htaccess files, if Apache's document root
+    # concurs with nginx's one
+    #
+    #location ~ /\.ht {
+    #    deny  all;
+    #}
+# }
+
